Splitting the C custom space into its own file, c_space.c/h Renaming check_grounding_space test to check_space
