+++ c/tests/CMakeLists.txt
@@ -1,15 +1,16 @@
 set(TEST_SOURCES
     test.c
     int_gnd.c
+    c_space.c
     util.c
 )
 add_executable(check_atom check_atom.c ${TEST_SOURCES})
 target_link_libraries(check_atom hyperonc CONAN_PKG::libcheck)
 add_test(NAME check_atom COMMAND check_atom)
 
-add_executable(check_grounding_space check_grounding_space.c ${TEST_SOURCES})
-target_link_libraries(check_grounding_space hyperonc CONAN_PKG::libcheck)
-add_test(NAME check_grounding_space COMMAND check_grounding_space)
+add_executable(check_space check_space.c ${TEST_SOURCES})
+target_link_libraries(check_space hyperonc CONAN_PKG::libcheck)
+add_test(NAME check_space COMMAND check_space)
 
 add_executable(check_sexpr_parser check_sexpr_parser.c ${TEST_SOURCES})
 target_link_libraries(check_sexpr_parser hyperonc CONAN_PKG::libcheck)
